one step closer; refactored NBT handling

+++ src/main/java/net/foxelfire/longer_boats/screen/LongBoatScreen.java
@@ -65,7 +65,7 @@ public class LongBoatScreen extends HandledScreen<LongBoatScreenHandler>{
         renderBackground(context, mouseX, mouseY, delta);
         super.render(context, mouseX, mouseY, delta);
         int magicXOffset = (width - backgroundWidth) / 2 + (int)(backgroundWidth/2.4);
-        int magicYOffset = (int)((height - backgroundHeight-32)/2 + backgroundHeight/7.75);
+        int magicYOffset = (int)((double) (height - backgroundHeight - 32) /2 + backgroundHeight/7.75);
         context.drawText(textRenderer, (handler.getCurrentTab()+1) + " / " + handler.entity.getNumberOfChests(), magicXOffset, magicYOffset, 0x303030, false);
         drawMouseoverTooltip(context, mouseX, mouseY);
     }
